create utils mapIcon
