+++ web/src/pages/OrphanagesMap.tsx
@@ -2,22 +2,12 @@ import React from 'react';
 import { Link } from 'react-router-dom';
 import { FiPlus, FiArrowRight } from 'react-icons/fi';
 import { MapContainer, TileLayer, Marker,Popup } from 'react-leaflet';
-import Leaflet from 'leaflet';
-
-import 'leaflet/dist/leaflet.css';
 
 import mapMarkerImg from '../images/map-marker.svg';
+import mapIcon from '../utils/mapIcon';
 
 import '../styles/pages/orphanages-map.css';
 
-const mapIcon = Leaflet.icon({
-    iconUrl: mapMarkerImg,
-
-    iconSize: [58,68],
-    iconAnchor: [29,68],
-
-    popupAnchor: [170,2],
-});
 
 function OrphanagesMap() {
     return(
